+++ rules/new-for-builtins.js
@@ -1,41 +1,10 @@
 'use strict';
 const getDocumentationUrl = require('./utils/get-documentation-url');
+const builtins = require('./utils/builtins');
 const isShadowed = require('./utils/is-shadowed');
 
-const enforceNew = new Set([
-	'Object',
-	'Array',
-	'ArrayBuffer',
-	'BigInt64Array',
-	'BigUint64Array',
-	'DataView',
-	'Date',
-	'Error',
-	'Float32Array',
-	'Float64Array',
-	'Function',
-	'Int8Array',
-	'Int16Array',
-	'Int32Array',
-	'Map',
-	'WeakMap',
-	'Set',
-	'WeakSet',
-	'Promise',
-	'RegExp',
-	'Uint8Array',
-	'Uint16Array',
-	'Uint32Array',
-	'Uint8ClampedArray'
-]);
-
-const disallowNew = new Set([
-	'BigInt',
-	'Boolean',
-	'Number',
-	'String',
-	'Symbol'
-]);
+const enforceNew = new Set(builtins.enforceNew);
+const disallowNew = new Set(builtins.disallowNew);
 
 const create = context => {
 	return {
